+++ clown_sort/sort_selector.py
@@ -2,21 +2,17 @@
 Open a GUI window to allow manual name / select.
 TODO: rename to something more appropriate
 """
-import shutil
 import sys
 from os import path, remove
-from subprocess import run
-
-from rich.panel import Panel
-from rich.text import Text
 
 from clown_sort.config import Config
 from clown_sort.filename_extractor import FilenameExtractor
-from clown_sort.util.dict_helper import get_dict_key_by_value
 from clown_sort.util.logging import log
 from clown_sort.util.rich_helper import bullet_text, console, indented_bullet
+from clown_sort.util.string_helper import is_empty
 
 RADIO_COLS = 11
+SELECT_SIZE = 45
 DELETE = 'Delete'
 OK = 'Move'
 OPEN = 'Preview Image'
@@ -25,72 +21,74 @@ EXIT = 'Exit'
 
 
 def process_file_with_popup(image: 'ImageFile') -> None:
-    import PySimpleGUI as sg
-    file_msg = Text("Processing: '").append(str(image.file_path), style='cyan reverse').append("'...")
-    console.print(Panel(file_msg, expand=False))
-    extracted_text = ' '.join((image.extracted_text() or '').splitlines())
-    log.info(f"OCR Text: {extracted_text} ({len(extracted_text)} chars)")
+    # Do the import here so as to allow usage without installing PySimpleGUI
+    import PySimpleGUI as psg
+    psg.theme('SystemDefault1')
     suggested_filename = FilenameExtractor(image).filename()
-    input = sg.Input(suggested_filename, size=(len(suggested_filename), 1))
+    sort_dirs = [path.basename(dir) for dir in Config.get_sort_dirs()]
+    max_dirname_length = max([len(dir) for dir in sort_dirs])
 
     layout = [
-        [sg.Image(data=image.image_bytes(), key="-IMAGE-")],
-        #[sg.Text(image.extracted_text())],
-        [sg.Text("Enter file name:")],
-        [input]
-    ] + list(_subdir_radio_select_columns(sg)) + \
-        [[sg.Button(OK, bind_return_key=True), sg.Button(DELETE), sg.Button(OPEN), sg.Button(SKIP), sg.Button(EXIT)]]
+        [psg.Column([[psg.Image(data=image.thumbnail_bytes(), key="-IMAGE-")]], justification='center')],
+        [psg.HSep()],
+        [psg.Text("Choose Filename:")],
+        [psg.Input(suggested_filename, size=(len(suggested_filename), 1))],# font=("Courier New", 12))],
+        [
+            psg.Text(f"Choose Directory:"),
+            psg.Combo(sort_dirs, size=(max_dirname_length, SELECT_SIZE)),
+            psg.Text(f"(Enter custom text to create new directory. If no directory is chosen file will be copied to '{Config.sorted_screenshots_dir}'.)")
+        ],
+        [
+            psg.Button(OK, bind_return_key=True),
+            psg.Button(DELETE),
+            psg.Button(OPEN),
+            psg.Button(SKIP),
+            psg.Button(EXIT)
+        ]
+    ]
 
-    window = sg.Window(image.basename, layout)
+    window = psg.Window(image.basename, layout)
 
     # Event Loop
     while True:
         event, values = window.Read()
 
+        if event == OPEN:
+            image.preview()
+            continue
+
+        window.close()
+
         if event == DELETE:
             log.warning(f"Deleting '{image.file_path}'")
             remove(image.file_path)
-            window.close()
             return
-        elif event == OPEN:
-            log.info(f"Opening '{image.file_path}'")
-            run(['open', image.file_path])
-        elif event == SKIP:
-            window.close()
+        elif event == SKIP or event == psg.WIN_CLOSED:
             return
         elif event == EXIT:
-            window.close()
             sys.exit()
         elif event == OK:
             break
 
-    window.close()
+    log.debug(f"All values: {values}")
     chosen_filename = values[0]
+    new_subdir = values[1]
+    destination_dir = Config.sorted_screenshots_dir.joinpath(new_subdir)
 
-    if chosen_filename is None or len(chosen_filename) == 0:
+    if is_empty(chosen_filename):
         raise ValueError("Filename can't be blank!")
 
-    try:
-        new_dir = get_dict_key_by_value(values, True)
-    except ValueError:
-        new_dir = Config.sorted_screenshots_dir
+    if not destination_dir.exists():
+        result = psg.popup_yes_no(f"Subdir '{new_subdir}' doesn't exist. Create?",  title="Unknown Subdirectory")
 
-    new_filename = path.join(new_dir, chosen_filename)
-    log.info(f"Chosen Filename: '{chosen_filename}'\nDirectory: '{new_dir}'\nNew file: '{new_filename}'\nEvent: {event}\n")
-    log.debug(f"All values: {values}")
+        if result == 'Yes' and not Config.dry_run:
+            log.info(f"Creating directory '{new_subdir}'...")
+            destination_dir.mkdir()
+        else:
+            console.print(bullet_text(f"Directory not found. Skipping '{image.file_path}'..."))
+            return
+
+    new_filename = destination_dir.joinpath(chosen_filename)
+    log.info(f"Chosen Filename: '{chosen_filename}'\nSubdir: '{new_subdir}'\nNew file: '{new_filename}'\nEvent: {event}\n")
     console.print(bullet_text(f"Moving '{image.file_path}' to '{new_filename}'..."))
-
-    if Config.dry_run:
-        console.print(indented_bullet("Dry run so not moving..."), style='dim')
-    else:
-        shutil.move(image.file_path, new_filename)
-
-
-def _subdir_radio_select_columns(sg):
-    dirs = Config.get_sort_dirs()
-
-    for i in range(0, len(dirs), RADIO_COLS):
-        yield [
-            sg.Radio(path.basename(dir), "SORTDIR_RADIO", default=False, key=dir)
-            for dir in dirs[i: i + RADIO_COLS]
-        ]
+    image.copy_file_to_sorted_dir(new_filename)